pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rialto-mcp"
version = "0.1.0"
description = "Rialto Investment MCP Server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "mcp>=1.26.0",
    "pydantic>=2.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["backend*", "gateway*"]

[project.scripts]
rialto = "backend.main:run"
rialto-gateway = "gateway.app:run"
